refactor(style): StyleReferencePanelを独立コンポーネント化+テスト、alt二重読み上げ修正
- Geminiレビュー対応: ズームボタン内の Image を alt="" (装飾)に。ボタンの
  aria-label が名前を担うため二重読み上げを防ぐ
- StyleReferencePanel を StylePageClient から別ファイルへ抽出(単体テスト可能に)
- ライトボックスのズーム開閉/alt構造の3ケースのテストを追加(codecov対応)
- 装飾alt化に伴い style-page-client.test の Style画像アサートを src 存在確認へ更新
