+++ features/style/components/StylePageClient.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { StyleProviderCredit } from "@/features/style/components/StyleProviderCredit";
 import {
@@ -12,9 +11,9 @@ import {
   type MouseEvent as ReactMouseEvent,
 } from "react";
 import { useLocale, useTranslations } from "next-intl";
-import { Maximize2, Minimize2, Share2, ZoomIn } from "lucide-react";
+import { Maximize2, Minimize2, Share2 } from "lucide-react";
 import { Card } from "@/components/ui/card";
-import { Dialog, DialogContent, DialogTitle } from "@/components/ui/dialog";
+import { StyleReferencePanel } from "@/features/style/components/StyleReferencePanel";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
@@ -224,124 +223,9 @@ function resolveInitialSelectedPresetId(
   return presets[0]?.id ?? "";
 }
 
-function StyleReferencePanel({
-  label,
-  imageSrc,
-  imageAlt,
-  className,
-  collapsed = false,
-  aspectRatio,
-  tooltip,
-  providerOverlay,
-}: {
-  label: string;
-  imageSrc: string;
-  imageAlt: string;
-  className?: string;
-  collapsed?: boolean;
-  aspectRatio?: number;
-  /**
-   * 画像コンテナの右上に絶対配置で重ねる任意の要素 (主に LabelInfoTooltip の `?` を想定)。
-   * preset カテゴリの user_guidance を画像内に表示するために使う。
-   */
-  tooltip?: React.ReactNode;
-  /**
-   * 画像コンテナの左下に絶対配置で重ねる任意の要素 (提供者クレジットを想定)。
-   * タップで提供者プロフィールへ遷移できる。
-   */
-  providerOverlay?: React.ReactNode;
-}) {
-  const t = useTranslations("style");
-  const [zoomed, setZoomed] = useState(false);
-  const ar = aspectRatio ?? 1;
-
-  return (
-    <div className={className ?? "space-y-3"}>
-      <Label
-        className={
-          collapsed
-            ? "text-xs font-medium leading-none"
-            : "text-base font-medium"
-        }
-      >
-        {label}
-      </Label>
-      <Card className="overflow-hidden p-0">
-        <div
-          className="relative bg-slate-100"
-          style={{ aspectRatio: String(ar) }}
-        >
-          {/* 画像タップで全画面表示(横長サンプルが小さくて見えない問題の対策)。
-              ツールチップ/提供者クレジットは z-20 で上に載せ、各自のタップを維持する。 */}
-          <button
-            type="button"
-            onClick={() => setZoomed(true)}
-            className="absolute inset-0 z-0 cursor-zoom-in focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-slate-500"
-            aria-label={t("styleImageZoomAria")}
-          >
-            <Image
-              src={imageSrc}
-              alt={imageAlt}
-              fill
-              sizes="(max-width: 768px) 100vw, 50vw"
-              className="object-cover"
-              priority
-            />
-          </button>
-          {/* タップで拡大できる合図。四隅で唯一どの要素とも競合しない左上に置く
-              (右上=ツールチップ / 左下=提供者クレジットは長い名前で max-w-[180px] まで伸びる)。 */}
-          <span
-            className={`pointer-events-none absolute z-10 flex items-center justify-center rounded-full bg-black/45 text-white ${
-              collapsed ? "left-1 top-1 h-5 w-5" : "left-2 top-2 h-7 w-7"
-            }`}
-            aria-hidden="true"
-          >
-            <ZoomIn className={collapsed ? "h-3 w-3" : "h-4 w-4"} />
-          </span>
-          {tooltip ? (
-            <div
-              className={`absolute z-20 ${collapsed ? "right-1 top-1" : "right-2 top-2"}`}
-            >
-              {tooltip}
-            </div>
-          ) : null}
-          {providerOverlay ? (
-            <div
-              className={`absolute z-20 ${collapsed ? "bottom-1 left-1" : "bottom-2 left-2"}`}
-            >
-              {providerOverlay}
-            </div>
-          ) : null}
-        </div>
-      </Card>
-
-      <Dialog open={zoomed} onOpenChange={setZoomed}>
-        <DialogContent
-          showCloseButton
-          className="border-0 bg-transparent p-0 shadow-none sm:max-w-[95vw]"
-          style={{ width: `min(95vw, calc(85vh * ${ar}))` }}
-        >
-          <DialogTitle className="sr-only">{label}</DialogTitle>
-          <div
-            className="relative w-full overflow-hidden rounded-lg bg-slate-900"
-            style={{ aspectRatio: String(ar), maxHeight: "85vh" }}
-          >
-            <Image
-              src={imageSrc}
-              alt={imageAlt}
-              fill
-              sizes="95vw"
-              className="object-contain"
-            />
-          </div>
-        </DialogContent>
-      </Dialog>
-    </div>
-  );
-}
-
-// 旧 StyleResultPanel は features/generation/components/GenerationResultPanel.tsx
-// に抽出した。
+// StyleReferencePanel は features/style/components/StyleReferencePanel.tsx に抽出した
+// (画像タップで全画面ライトボックスを開く。単体テストのため独立コンポーネント化)。
+// 旧 StyleResultPanel は features/generation/components/GenerationResultPanel.tsx に抽出した。
 
 async function fetchStyleRateLimitStatus(): Promise<StyleRateLimitStatusState | null> {
   const response = await fetch("/style/rate-limit-status", {
